pyproject.toml
[tool.poetry]
name = "riskmap"
version = "0.1"
description = "Mapping red team tools to risk frameworks"
authors = ["Leron Gray <daddycocoaman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
stix2 = "^2.1.0"
loguru = "^0.5.3"
prettytable = "^2.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
typer = {version = "^0.3.2", extras = ["all"]}
azure-identity = "^1.5.0"
azure-mgmt-resource = "^15.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
